+++ test-data-regression/test_netlists_sample_doe_function_.yml
@@ -1179,7 +1179,7 @@ placements:
   mzis:
     mirror: 0
     rotation: 0
-    x: 1458.698
+    x: 1458.699
     y: 0.0
   rings:
     mirror: 0
@@ -1345,19 +1345,19 @@ warnings:
         - 1013.001
       - - 1277.853
         - 1013.001
-      - - 1430.798
-        - 78.676
-      - - 1557.798
-        - 78.676
-      - - 1303.798
-        - 78.676
-      - - 1684.798
-        - 78.676
-      - - 1430.798
+      - - 1430.799
+        - 78.677
+      - - 1557.799
+        - 78.677
+      - - 1303.799
+        - 78.677
+      - - 1684.799
+        - 78.677
+      - - 1430.799
         - -100.3
-      - - 1557.798
+      - - 1557.799
         - -100.3
-      - - 1303.798
+      - - 1303.799
         - -100.3
-      - - 1684.798
+      - - 1684.799
         - -100.3
