+++ test-data-regression/test_netlists_sample_rotation_.yml
@@ -49,4 +49,4 @@ warnings:
       - - 8.0
         - 1.0
       - - 6.0
-        - 0.0
+        - -0.0
