+++ test-data-regression/test_settings_add_fiber_array_optical_south_electrical_north_.yml
@@ -3,7 +3,7 @@ ports:
   e4_1_1:
     center:
     - 85.5
-    - 129.655
+    - 129.656
     layer:
     - 49
     - 0
@@ -15,7 +15,7 @@ ports:
   e4_1_2:
     center:
     - 185.5
-    - 129.655
+    - 129.656
     layer:
     - 49
     - 0
