+++ test-data-regression/test_settings_cutback_bend180_.yml
@@ -14,7 +14,7 @@ ports:
     width: 0.5
   o2:
     center:
-    - 203.282
+    - 203.28
     - 0.0
     layer:
     - 1
