+++ test-data-regression/test_settings_disk_heater_.yml
@@ -3,7 +3,7 @@ ports:
   e1:
     center:
     - -22.551
-    - 0.25
+    - 0.251
     layer:
     - 49
     - 0
@@ -15,7 +15,7 @@ ports:
   e2:
     center:
     - 33.251
-    - 0.25
+    - 0.251
     layer:
     - 49
     - 0
