+++ test-data-regression/test_settings_straight_heater_meander_doped_.yml
@@ -51,7 +51,7 @@ ports:
   o1:
     center:
     - -25.0
-    - -0.0
+    - 0.0
     layer:
     - 1
     - 0
